requirements.txt
# Core dependencies
PyMuPDF
openai
python-dotenv
requests
pydantic
tiktoken

# Interface
colorama
tqdm

# Authentication
PyJWT
